improvement: Partition back pressure from encoder space required
As this is only calculated at the end of a batch cycle, and not continuously,
it may be too late, and offsets have to be dropped. But it’s relatively
simple and get’s us a lot of the way to the continuous (correct) solution).

Bitset test and fixes

Test failed messages that are blocking the partition can always retry, test pessimistic partition blocking (buffer of available payload space)