fix: Allow any record to be attempted for reprocessing, not just retries
Only allowing retries is naive as there may be records which are already
represented in the payload encoding which, if completed, may reduce it's size.

The encoders must encode all state up to the high succeeded offset, as any beyond that are
by definition incomplete and can be ignored. Any record who's offset is below this, will
be contributing to the size of the payload. Allowing them to attempt to be completed may
reduce the payload size.