Idea: use send call back to trigger controller to wake up, also cache list of expected records to produce, and count them down as they finish, when they do, onSuccess each wc.
Issue is that our api doesn't connect in records without records in the batching API, so we don't know which out record corresponds with which WC. We only know when our expected sent count reaches actual sent count.

There is still some questions to answer here with how this interacts with transactions, but as long as the transaction doesn't commit until everything's complete. And new records are blocked, and retrying any failures doesn't result in dupes which it shouldn't.