START: fixes #419 NoSuchElementException during race condition in PartitionState
In busy situations, there is a race condition in PartitionState due to it being edited by the Controller, and read by the BrokerPoller.

Because the check and subsequent read of the incompleteOffsets collection are two different operations, it can initially appear as though it has elements, but when we try to retrieve an element, it can then be empty, which throws a NoSuchElementException.

Thanks to a very elegant fix suggested by @shaileshkulkarni, we can combine the check and retrieval by using the thread safe #ceiling method, which can either returns the #first element, or null if empty.

This is as opposed to the more brute force solution which would require using locks.

This, among other things, will be made redundant once #200 (a shared nothing architecture) is merged.