START: Lookup committed() data for partition load directly from consumer owner
Instead of looking it up later, and having to access to consumer. Previously the poller would have the onAssigned methods trigger from the consumer, which would in turn run through the shared WorkManager, and then have to call #committed() on the consumer. Now, the #committed() data is sent along with the #onAssigned() message to the controller.