fix: After new performance fix PR#530 merges - corner case could cause out of order processing (#534)
Under unrealistically high load with no-op processing, broker poller unblocking a partition could cause ProcessingShard to skip forward in its entries and take work out of order.

This was discovered when fixing a synthetic high performance benchmark, after PR#530 (O(n) algo was fixed to O(1)), creating the state for the race condition to appear. Probably could not happen without the fix, as it's related to the performance of certain parts of the system.